Filtering Temperature

+++ Day4-Task2/Classes/Weather.cs
@@ -11,10 +11,10 @@ namespace Day4_Task2.Classes
         public string Country { get; set; }
         public int ZipCode { get; set; }
         public string City { get; set; }
-        public decimal Temperature { get; set; }
+        public int Temperature { get; set; }
         public decimal WindSpeed { get; set; }
 
-        public Weather(string country,int zipcode, string city, decimal temperature, decimal windspeed)
+        public Weather(string country,int zipcode, string city, int temperature, decimal windspeed)
         {
             Country = country;
             ZipCode = zipcode;
